Remove fasd again
Together with an alias ...='cd ../..' fasd causes weird behaviour when
using an anonymous function like '() { true }'.
The reason seems to be that the _fasd_preexec function operates on $2
which is a reduced form of the command. The reduced form hides function
bodies as '...'
I didn't dig further, but it seems this leads to fasd executing '...' as
a preexec hook which is aliased to 'cd ../..' on my machine, so on every
command invocation, zsh moves two folders up.